@@ -1,11 +1,11 @@
 name: downloads_path_provider_28
 description: Flutter plugin to get the download directory for Android modified for sdk version 28+. Forked from Daniel MF.
-version: 0.1.2
+version: 0.0.1
 homepage: https://github.com/chitgoks/downloads_path_provider
 
 environment:
-  flutter: ">=1.12.0"
-  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
