@@ -1,10 +1,9 @@
+import 'dart:async';
 import 'dart:io';
 
-import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:downloads_path_provider_28/downloads_path_provider_28.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 void main() => runApp(new MyApp());
 
@@ -14,7 +13,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  Directory _downloadsDirectory;
+  Directory? _downloadsDirectory;
 
   @override
   void initState() {
@@ -24,7 +23,7 @@ class _MyAppState extends State<MyApp> {
 
   // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initDownloadsDirectoryState() async {
-    Directory downloadsDirectory;
+    Directory? downloadsDirectory;
     // Platform messages may fail, so we use a try/catch PlatformException.
     try {
       downloadsDirectory = await DownloadsPathProvider.downloadsDirectory;
@@ -52,7 +51,7 @@ class _MyAppState extends State<MyApp> {
         body: new Center(
           child: new Text(
             _downloadsDirectory != null
-                ? 'Downloads directory: ${_downloadsDirectory.path}\n'
+                ? 'Downloads directory: ${_downloadsDirectory!.path}\n'
                 : 'Could not get the downloads directory',
           ),
         ),
